refactor(nitro): split out clearVueCurrentInstance
