+++ packages/nitro-server/src/runtime/handlers/island.ts
@@ -13,7 +13,8 @@ import { traceAsync } from '#app/internal/tracing'
 // @ts-expect-error virtual file
 import { tracingChannelNuxt } from '#internal/nuxt.config.mjs'
 import { islandCache, islandPropCache } from '../utils/cache'
-import { clearVueCurrentInstance, createSSRContext } from '../utils/renderer/app'
+import { createSSRContext } from '../utils/renderer/app'
+import { clearVueCurrentInstance } from '../utils/renderer/clear-instance'
 import { getSSRRenderer } from '../utils/renderer/build-files'
 import { renderInlineStyles } from '../utils/renderer/inline-styles'
 import { getClientIslandResponse, getServerComponentHTML, getSlotIslandResponse } from '../utils/renderer/islands'
